requirement.txt
pycryptodome==3.23.0
pyperclip==1.11.0
